backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
websockets==12.0
faster-whisper==1.2.1
ffmpeg-python==0.2.0
av>=12.0.0
pyyaml==6.0.1
